\name{brca}
\alias{brca}
\docType{data}
\title{
Modified TCGA Breast Cancer data
}
\description{
Modified TCGA Breast Cancer data.  
}
\usage{data(brca)}
\format{
  Three data frames, with columns as standardized miRNA, gene, and protein expressions. One vector with the two classes of the samples (tumor or normal tissue)
}
\keyword{datasets}
