\name{iDINGO-package}
\alias{iDINGO-package}
\alias{iDINGO}
\docType{package}
\title{
iDINGO: Integrative Differential Network Analysis in Genomics
}
\description{
This packages jointly estimates group specific partial correlations in a multi-level/platform data set.
}
\details{
This packages jointly estimates group specific partial correlations in a multi-level/platform data set, considering the directionality of effects between platforms using a chain graph model.
}
\author{
Min Jin Ha, Caleb Class, Veerabhadran Baladandayuthapani, and Kim-Anh Do
}
\keyword{ package }