\name{scaledMat}
\alias{scaledMat}
\title{
scale a square matrix
}
\description{
scale a square matrix to have unit diagonal elements.
}
\usage{
scaledMat(x)
}
\arguments{
  \item{x}{
   a square matrix with positive diagonal elements
}
}
\value{
  scaled matrix of x
}
\author{
Min Jin Ha mjha@mdanderson.org
}