\name{trans.Fisher}
\alias{trans.Fisher}

\title{
Fisher's Z-transformation
}
\description{
Fisher's Z-transformation of (partial) correlation.
}

\arguments{
  \item{x}{
    a vector having entries between -1 and 1
}
}
\value{
    Fisher's Z-transformed values
}
\author{
    Min Jin HA mjha@mdanderson.org	
}
